+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog of RorVsWild
 
+
+## Unreleased
+
+* Compress data before sending
+
 ## 1.10.1 (2025-12-24)
 
 * Fix read_last_lines for local widget